#137 - Separate the projects-table container into a container
