+++ src/components/projects-table/projects-table.module.css
@@ -1,0 +1,8 @@
+/*
+ * This file is part of NER's PM Dashboard and licensed under GNU AGPLv3.
+ * See the LICENSE file in the repository root folder for details.
+ */
+
+.table {
+  padding: 0em 2em 0em 2em;
+}
