+++ src/containers/projects-table/projects-table.test.tsx
@@ -109,100 +109,4 @@ describe('projects table component', () => {
       expectedWbsOrder.reverse()
     );
   });
-
-  it('handles sorting and reverse sorting the table by project name', async () => {
-    render(<ProjectsTable />);
-    await waitFor(() => screen.getByText(wbsPipe(exampleProject1.wbsNum)));
-
-    const column: string = 'Name';
-    const expectedWbsOrder: string[] = [
-      exampleProject3,
-      exampleProject2,
-      exampleProject1,
-      exampleProject4,
-      exampleProject5
-    ].map((prj: Project) => wbsPipe(prj.wbsNum));
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsDesc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsDesc.map((ele: HTMLElement) => ele.innerHTML)).toEqual(
-      expectedWbsOrder.reverse()
-    );
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsAsc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsAsc.map((ele: HTMLElement) => ele.innerHTML)).toEqual(expectedWbsOrder.reverse());
-  });
-
-  it('handles sorting and reverse sorting the table by project lead', async () => {
-    render(<ProjectsTable />);
-    await waitFor(() => screen.getByText(wbsPipe(exampleProject1.wbsNum)));
-
-    const column: string = 'Project Lead';
-    const expectedWbsOrder: string[] = [
-      exampleProject1,
-      exampleProject2,
-      exampleProject5,
-      exampleProject3,
-      exampleProject4
-    ].map((prj: Project) => wbsPipe(prj.wbsNum));
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsDesc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsDesc.map((ele: HTMLElement) => ele.innerHTML)).not.toEqual(expectedWbsOrder);
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsAsc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsAsc.map((ele: HTMLElement) => ele.innerHTML)).toEqual(expectedWbsOrder);
-  });
-
-  it('handles sorting and reverse sorting the table by project manager', async () => {
-    render(<ProjectsTable />);
-    await waitFor(() => screen.getByText(wbsPipe(exampleProject1.wbsNum)));
-
-    const column: string = 'Project Manager';
-    const expectedWbsOrder: string[] = [
-      exampleProject1,
-      exampleProject4,
-      exampleProject2,
-      exampleProject3,
-      exampleProject5
-    ].map((prj: Project) => wbsPipe(prj.wbsNum));
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsDesc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsDesc.map((ele: HTMLElement) => ele.innerHTML)).not.toEqual(expectedWbsOrder);
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsAsc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsAsc.map((ele: HTMLElement) => ele.innerHTML)).toEqual(expectedWbsOrder);
-  });
-
-  it('handles sorting and reverse sorting the table by duration', async () => {
-    render(<ProjectsTable />);
-    await waitFor(() => screen.getByText(wbsPipe(exampleProject1.wbsNum)));
-
-    const column: string = 'Duration';
-    const expectedWbsOrder: string[] = [
-      exampleProject2,
-      exampleProject5,
-      exampleProject3,
-      exampleProject4,
-      exampleProject1
-    ].map((prj: Project) => wbsPipe(prj.wbsNum));
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsDesc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsDesc.map((ele: HTMLElement) => ele.innerHTML)).toEqual(
-      expectedWbsOrder.reverse()
-    );
-
-    fireEvent.click(screen.getByText(column));
-    const wbsNumsAsc: HTMLElement[] = await screen.findAllByText(wbsRegex);
-    expect(wbsNumsAsc.map((ele: HTMLElement) => ele.innerHTML)).toEqual(expectedWbsOrder.reverse());
-  });
-
-  // test for a basic normal case
-
-  // https://testing-library.com/docs/react-testing-library/example-intro
 });
