+++ src/containers/projects-table/projects-table.tsx
@@ -3,34 +3,17 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { useState, useEffect } from 'react';
-import { useHistory } from 'react-router-dom';
-import BootstrapTable, {
-  ColumnDescription,
-  RowEventHandlerProps,
-  SortOrder
-} from 'react-bootstrap-table-next';
-import { Project, WorkPackage } from 'utils';
+import { useEffect, useState } from 'react';
 import { AxiosResponse } from 'axios';
+import { Project, WorkPackage } from 'utils';
 import { apiFetch } from '../../shared/axios';
-import { fullNamePipe, wbsPipe, weeksPipe } from '../../shared/pipes';
-import styles from './projects-table.module.css';
+import { weeksPipe, fullNamePipe, wbsPipe } from '../../shared/pipes';
+import PrjsTable from '../../components/projects-table/projects-table'; // Directly rename the default import
+import { DisplayProject } from '../../components/projects-table/projects-table';
+import './projects-table.module.css';
 
-interface DisplayProject {
-  wbsNum: string;
-  name: string;
-  projectLead: string;
-  projectManager: string;
-  duration: string;
-}
-
-/**
- * Interactive table for fetching and displaying all projects data.
- */
-const ProjectsTable: React.FC = () => {
-  const history = useHistory();
-  const initial: DisplayProject[] = [];
-  const [allProjects, setAllProjects] = useState(initial); // store projects data
+const ProjectsTable: React.FC = () => {  
+  const [allProjects, setAllProjects] = useState<DisplayProject[]>([]); // store projects data
 
   // Transforms given project data and sets local state
   const updateData: Function = (response: AxiosResponse) => {
@@ -67,49 +50,8 @@ const ProjectsTable: React.FC = () => {
     };
   }, []);
 
-  // Configures display options for all data columns
-  // TODO: Sort by wbsNum means 1.1.0 > 1.12.0 > 1.2.0, but desired is 1.1.0 > 1.2.0 > 1.12.0
-  // TODO: Sort by duration means 12 > 2 > 4 > 5 > 9, but desired is 12 > 9 > 5 > 4 > 2
-  const columns: ColumnDescription[] = [
-    { dataField: 'wbsNum', text: 'WBS #', align: 'center', sort: true },
-    { dataField: 'name', text: 'Name', align: 'left', sort: true },
-    { dataField: 'projectLead', text: 'Project Lead', align: 'left', sort: true },
-    { dataField: 'projectManager', text: 'Project Manager', align: 'left', sort: true },
-    { dataField: 'duration', text: 'Duration', align: 'center', sort: true }
-  ];
-
-  const defaultSort: [{ dataField: any; order: SortOrder }] = [
-    {
-      dataField: 'wbsNum',
-      order: 'asc'
-    }
-  ];
-
-  // define what happens during various row events
-  const rowEvents: RowEventHandlerProps = {
-    onClick: (e, row, rowIndex) => {
-      history.push(`/projects/${row.wbsNum}`);
-    }
-  };
-
-  return (
-    <>
-      <h3>This is the Projects Table container</h3>
-      <BootstrapTable
-        striped
-        hover
-        condensed
-        wrapperClasses={styles.table}
-        bootstrap4={true}
-        keyField="wbsNum"
-        data={allProjects}
-        columns={columns}
-        defaultSorted={defaultSort}
-        rowEvents={rowEvents}
-        noDataIndication="No Projects to Display"
-      />
-    </>
-  );
+  return <PrjsTable allProjects={allProjects} />;
 };
+
 
 export default ProjectsTable;
